Question downloading and deserializing working, but not yet displaying properly.
